requirements.txt
asttokens==2.0.8
backcall==0.2.0
beautifulsoup4==4.11.1
black==22.8.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
decorator==5.1.1
executing==1.0.0
idna==3.4
ipdb==0.13.9
ipython==8.5.0
jedi==0.18.1
Jinja2==3.1.2
jinja2-markdown==0.0.3
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mypy-extensions==0.4.3
parso==0.8.3
pathspec==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.2
prompt-toolkit==3.0.31
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
pymediawiki==0.7.2
python-frontmatter==1.0.0
requests==2.28.1
rfeed==1.1.1
six==1.16.0
soupsieve==2.3.2.post1
stack-data==0.5.0
toml==0.10.2
tomli==2.0.1
traitlets==5.4.0
urllib3==1.26.12
wcwidth==0.2.5
wikipedia==1.4.0
